finalize featured item creation and edit, fix product edit bug

+++ backend/delivery/app/Http/Controllers/Admin/ProductsController.php
@@ -80,9 +80,11 @@ class ProductsController {
 
             $product->save();
             
-            // all changes succeeded delete old image
-            if ($oldImagePath) {
-                Storage::disk('public')->delete($oldImagePath);
+            if ($hasImageField) {
+                // all changes succeeded delete old image
+                if ($oldImagePath) {
+                    Storage::disk('public')->delete($oldImagePath);
+                }
             }
         });
     }
